Transferred program functionality to individual functions from main script
The script still declares some important variables, however, i created a functions file with all of the functionality such as 'send_email()' and started incorporating a new functionality called 'send_text().'

'send_text()' is currently incomplete, however, I hope to incorporate this into the invoice send to clients.

Then, the clients will receive an email and text delivery of invoice payload.

Also, created 'test-branch' to house beta files that are not officially adopted in the program.